refactor(load): move all load and progress into their packages
