+++ pkg/progress/progress.go
@@ -1,4 +1,4 @@
-package commands
+package progress
 
 import (
 	"context"
