+++ pkg/progress/progress_test.go
@@ -1,4 +1,4 @@
-package commands
+package progress
 
 import (
 	"context"
